Refactor ArtModels
